// src/app/styled-css/padding.style/padding.d.ts
import { Property } from 'csstype';

export interface PaddingStyleProps {
  padding?: Property.Padding;
  paddingXxl?: Property.Padding;
  paddingXl?: Property.Padding;
  paddingLg?: Property.Padding;
  paddingMd?: Property.Padding;
  paddingSm?: Property.Padding;
  paddingXs?: Property.Padding;

  paddingTop?: Property.Padding;
  paddingTopXxl?: Property.Padding;
  paddingTopXl?: Property.Padding;
  paddingTopLg?: Property.Padding;
  paddingTopMd?: Property.Padding;
  paddingTopSm?: Property.Padding;
  paddingTopXs?: Property.Padding;

  paddingBottom?: Property.Padding;
  paddingBottomXxl?: Property.Padding;
  paddingBottomXl?: Property.Padding;
  paddingBottomLg?: Property.Padding;
  paddingBottomMd?: Property.Padding;
  paddingBottomSm?: Property.Padding;
  paddingBottomXs?: Property.Padding;

  paddingRight?: Property.Padding;
  paddingRightXxl?: Property.Padding;
  paddingRightXl?: Property.Padding;
  paddingRightLg?: Property.Padding;
  paddingRightMd?: Property.Padding;
  paddingRightSm?: Property.Padding;
  paddingRightXs?: Property.Padding;

  paddingLeft?: Property.Padding;
  paddingLeftXxl?: Property.Padding;
  paddingLeftXl?: Property.Padding;
  paddingLeftLg?: Property.Padding;
  paddingLeftMd?: Property.Padding;
  paddingLeftSm?: Property.Padding;
  paddingLeftXs?: Property.Padding;

  paddingTopBottom?: Property.Padding;
  paddingTopBottomXxl?: Property.Padding;
  paddingTopBottomXl?: Property.Padding;
  paddingTopBottomLg?: Property.Padding;
  paddingTopBottomMd?: Property.Padding;
  paddingTopBottomSm?: Property.Padding;
  paddingTopBottomXs?: Property.Padding;

  paddingLeftRight?: Property.Padding;
  paddingLeftRightXxl?: Property.Padding;
  paddingLeftRightXl?: Property.Padding;
  paddingLeftRightLg?: Property.Padding;
  paddingLeftRightMd?: Property.Padding;
  paddingLeftRightSm?: Property.Padding;
  paddingLeftRightXs?: Property.Padding;
}